refactor(weather-widget): simplify renderer and effect composition
